Move fixDate to standard functions, add some tests for null path
