+++ library/global_functions.inc.php
@@ -257,6 +257,45 @@ function display_desc($desc)
     return $desc;
 }
 
+// Supported input date formats are:
+//   mm/dd/yyyy
+//   mm/dd/yy   (assumes 20yy for yy < 10, else 19yy)
+//   yyyy/mm/dd
+//   also mm-dd-yyyy, etc. and mm.dd.yyyy, etc.
+//
+function fixDate($date, $default = "0000-00-00")
+{
+    $fixed_date = $default;
+    $date = trim((string) $date);
+    if (preg_match("'^[0-9]{1,4}[/.-][0-9]{1,2}[/.-][0-9]{1,4}$'", $date)) {
+        $dmy = preg_split("'[/.-]'", $date);
+        if ($dmy[0] > 99) {
+            $fixed_date = sprintf("%04u-%02u-%02u", $dmy[0], $dmy[1], $dmy[2]);
+        } else {
+            if ($dmy[0] != 0 || $dmy[1] != 0 || $dmy[2] != 0) {
+                if ($dmy[2] < 1000) {
+                    $dmy[2] += 1900;
+                }
+
+                if ($dmy[2] < 1910) {
+                    $dmy[2] += 100;
+                }
+            }
+            // Determine if MDY date format is used, preferring Date Display Format from
+            // global settings if it's not YMD, otherwise guessing from country code.
+            $using_mdy = empty(OEGlobalsBag::getInstance()->get('date_display_format')) ?
+                (OEGlobalsBag::getInstance()->getInt('phone_country_code') === 1) : (OEGlobalsBag::getInstance()->get('date_display_format') == 1);
+            if ($using_mdy) {
+                $fixed_date = sprintf("%04u-%02u-%02u", $dmy[2], $dmy[0], $dmy[1]);
+            } else {
+                $fixed_date = sprintf("%04u-%02u-%02u", $dmy[2], $dmy[1], $dmy[0]);
+            }
+        }
+    }
+
+    return $fixed_date;
+}
+
 /**
  * Format a money amount with decimals but no other decoration.
  *
